Update tests for auto download.
