+++ .gitignore
@@ -38,6 +38,7 @@ tmtagsHistory
 ######################
 exotic_ld_data/
 exotic_ld_data_original/
+tests/test_exotic_ld_data/
 *.p
 .ipynb_checkpoints
 docs/source/views/api/summary
